+++ .gitignore
@@ -7,6 +7,7 @@ plans/
 
 ## AI
 CLAUDE.md
+AGENTS.md
 
 ## User settings
 xcuserdata/
