+++ bitchat/Protocols/Packets.swift
@@ -1,0 +1,116 @@
+import Foundation
+
+// MARK: - Protocol TLV Packets
+
+struct AnnouncementPacket {
+    let nickname: String
+    let publicKey: Data
+
+    private enum TLVType: UInt8 {
+        case nickname = 0x01
+        case noisePublicKey = 0x02
+    }
+
+    func encode() -> Data? {
+        var data = Data()
+
+        // TLV for nickname
+        guard let nicknameData = nickname.data(using: .utf8), nicknameData.count <= 255 else { return nil }
+        data.append(TLVType.nickname.rawValue)
+        data.append(UInt8(nicknameData.count))
+        data.append(nicknameData)
+
+        // TLV for public key
+        guard publicKey.count <= 255 else { return nil }
+        data.append(TLVType.noisePublicKey.rawValue)
+        data.append(UInt8(publicKey.count))
+        data.append(publicKey)
+
+        return data
+    }
+
+    static func decode(from data: Data) -> AnnouncementPacket? {
+        var offset = 0
+        var nickname: String?
+        var publicKey: Data?
+
+        while offset + 2 <= data.count {
+            guard let type = TLVType(rawValue: data[offset]) else { return nil }
+            offset += 1
+
+            let length = Int(data[offset])
+            offset += 1
+
+            guard offset + length <= data.count else { return nil }
+            let value = data[offset..<offset + length]
+            offset += length
+
+            switch type {
+            case .nickname:
+                nickname = String(data: value, encoding: .utf8)
+            case .noisePublicKey:
+                publicKey = Data(value)
+            }
+        }
+
+        guard let nickname = nickname, let publicKey = publicKey else { return nil }
+        return AnnouncementPacket(nickname: nickname, publicKey: publicKey)
+    }
+}
+
+struct PrivateMessagePacket {
+    let messageID: String
+    let content: String
+
+    private enum TLVType: UInt8 {
+        case messageID = 0x00
+        case content = 0x01
+    }
+
+    func encode() -> Data? {
+        var data = Data()
+
+        // TLV for messageID
+        guard let messageIDData = messageID.data(using: .utf8), messageIDData.count <= 255 else { return nil }
+        data.append(TLVType.messageID.rawValue)
+        data.append(UInt8(messageIDData.count))
+        data.append(messageIDData)
+
+        // TLV for content
+        guard let contentData = content.data(using: .utf8), contentData.count <= 255 else { return nil }
+        data.append(TLVType.content.rawValue)
+        data.append(UInt8(contentData.count))
+        data.append(contentData)
+
+        return data
+    }
+
+    static func decode(from data: Data) -> PrivateMessagePacket? {
+        var offset = 0
+        var messageID: String?
+        var content: String?
+
+        while offset + 2 <= data.count {
+            guard let type = TLVType(rawValue: data[offset]) else { return nil }
+            offset += 1
+
+            let length = Int(data[offset])
+            offset += 1
+
+            guard offset + length <= data.count else { return nil }
+            let value = data[offset..<offset + length]
+            offset += length
+
+            switch type {
+            case .messageID:
+                messageID = String(data: value, encoding: .utf8)
+            case .content:
+                content = String(data: value, encoding: .utf8)
+            }
+        }
+
+        guard let messageID = messageID, let content = content else { return nil }
+        return PrivateMessagePacket(messageID: messageID, content: content)
+    }
+}
+
