+++ bitchat/Services/SimplifiedBluetoothService.swift
@@ -983,7 +983,7 @@ final class SimplifiedBluetoothService: NSObject {
         
         // Verify that the sender's derived ID from the announced public key matches the packet senderID
         // This helps detect relayed or spoofed announces. Only warn in release; assert in debug.
-        let derivedFromKey = Self.derivePeerID(fromPublicKey: announcement.publicKey)
+        let derivedFromKey = PeerIDUtils.derivePeerID(fromPublicKey: announcement.publicKey)
         if derivedFromKey != peerID {
             SecureLogger.log("⚠️ Announce sender mismatch: derived \(derivedFromKey.prefix(8))… vs packet \(peerID.prefix(8))…", category: SecureLogger.security, level: .warning)
             #if DEBUG
@@ -2082,12 +2082,6 @@ extension SimplifiedBluetoothService: CBPeripheralManagerDelegate {
 // MARK: - Advertising Builders & Alias Rotation
 
 extension SimplifiedBluetoothService {
-    static func derivePeerID(fromPublicKey publicKey: Data) -> String {
-        let digest = SHA256.hash(data: publicKey)
-        let hex = digest.map { String(format: "%02x", $0) }.joined()
-        return String(hex.prefix(16))
-    }
-    
     private func buildAdvertisementData() -> [String: Any] {
         let data: [String: Any] = [
             CBAdvertisementDataServiceUUIDsKey: [SimplifiedBluetoothService.serviceUUID]
@@ -2099,299 +2093,4 @@ extension SimplifiedBluetoothService {
     // No alias rotation or advertising restarts required.
 }
 
-// MARK: - Nostr Embedding Helpers
-
-extension SimplifiedBluetoothService {
-    /// Build a `bitchat1:` base64url-encoded BitChat packet carrying a private message
-    /// for transport over Nostr DMs. The payload is a plaintext typed NoisePayload
-    /// (no inner Noise encryption; NIP-17 provides transport-layer E2E).
-    func buildNostrEmbeddedPrivateMessageContent(content: String, to recipientPeerID: String, messageID: String) -> String? {
-        // TLV-encode the private message
-        let pm = PrivateMessagePacket(messageID: messageID, content: content)
-        guard let tlv = pm.encode() else { return nil }
-
-        // Prefix with NoisePayloadType
-        var payload = Data([NoisePayloadType.privateMessage.rawValue])
-        payload.append(tlv)
-
-        // Build BitChat packet (noiseEncrypted type used as a typed envelope)
-        // Determine correct 8-byte recipient ID (peerID) to embed
-        let recipientIDHex: String = {
-            if let maybeData = Data(hexString: recipientPeerID) {
-                if maybeData.count == 32 {
-                    // Treat as Noise static public key; derive peerID from fingerprint
-                    return Self.derivePeerID(fromPublicKey: maybeData)
-                } else if maybeData.count == 8 {
-                    // Already an 8-byte peer ID
-                    return recipientPeerID
-                }
-            }
-            // Fallback (should not happen): use myPeerID to avoid dropping
-            return recipientPeerID.count == 16 ? recipientPeerID : myPeerID
-        }()
-
-        let packet = BitchatPacket(
-            type: MessageType.noiseEncrypted.rawValue,
-            senderID: hexStringToData(myPeerID),
-            recipientID: hexStringToData(recipientIDHex),
-            timestamp: UInt64(Date().timeIntervalSince1970 * 1000),
-            payload: payload,
-            signature: nil,
-            ttl: messageTTL
-        )
-
-        guard let data = packet.toBinaryData() else { return nil }
-        return "bitchat1:" + Self.base64URLEncode(data)
-    }
-
-    /// Build a `bitchat1:` base64url-encoded BitChat packet carrying a delivery/read ack
-    /// for transport over Nostr DMs. Payload is plaintext typed NoisePayload.
-    func buildNostrEmbeddedAckContent(type: NoisePayloadType, messageID: String, to recipientPeerID: String) -> String? {
-        guard type == .delivered || type == .readReceipt else { return nil }
-
-        var payload = Data([type.rawValue])
-        payload.append(Data(messageID.utf8))
-
-        // Determine correct 8-byte recipient ID (peerID) to embed
-        let recipientIDHex: String = {
-            if let maybeData = Data(hexString: recipientPeerID) {
-                if maybeData.count == 32 {
-                    return Self.derivePeerID(fromPublicKey: maybeData)
-                } else if maybeData.count == 8 {
-                    return recipientPeerID
-                }
-            }
-            return recipientPeerID.count == 16 ? recipientPeerID : myPeerID
-        }()
-
-        let packet = BitchatPacket(
-            type: MessageType.noiseEncrypted.rawValue,
-            senderID: hexStringToData(myPeerID),
-            recipientID: hexStringToData(recipientIDHex),
-            timestamp: UInt64(Date().timeIntervalSince1970 * 1000),
-            payload: payload,
-            signature: nil,
-            ttl: messageTTL
-        )
-
-        guard let data = packet.toBinaryData() else { return nil }
-        return "bitchat1:" + Self.base64URLEncode(data)
-    }
-
-    /// Base64url encode without padding
-    private static func base64URLEncode(_ data: Data) -> String {
-        let b64 = data.base64EncodedString()
-        let urlSafe = b64
-            .replacingOccurrences(of: "+", with: "-")
-            .replacingOccurrences(of: "/", with: "_")
-            .replacingOccurrences(of: "=", with: "")
-        return urlSafe
-    }
-}
-
-// MARK: - Message Deduplicator
-
-/// Efficient message deduplication with time-based cleanup
-private class MessageDeduplicator {
-    private struct Entry {
-        let messageID: String
-        let timestamp: Date
-    }
-    
-    private var entries: [Entry] = []
-    private var lookup = Set<String>()
-    private let lock = NSLock()
-    private let maxAge: TimeInterval = 300  // 5 minutes
-    private let maxCount = 1000
-    
-    /// Check if message is duplicate and add if not
-    func isDuplicate(_ messageID: String) -> Bool {
-        lock.lock()
-        defer { lock.unlock() }
-        
-        // Clean old entries first
-        cleanupOldEntries()
-        
-        if lookup.contains(messageID) {
-            return true
-        }
-        
-        // Add new entry
-        entries.append(Entry(messageID: messageID, timestamp: Date()))
-        lookup.insert(messageID)
-        
-        // Efficient cleanup when over limit
-        if entries.count > maxCount {
-            // Remove oldest 100 entries
-            let toRemove = entries.prefix(100)
-            toRemove.forEach { lookup.remove($0.messageID) }
-            entries.removeFirst(100)
-        }
-        
-        return false
-    }
-    
-    /// Add an ID without checking (for announce-back tracking)
-    func markProcessed(_ messageID: String) {
-        lock.lock()
-        defer { lock.unlock() }
-        
-        if !lookup.contains(messageID) {
-            entries.append(Entry(messageID: messageID, timestamp: Date()))
-            lookup.insert(messageID)
-        }
-    }
-    
-    /// Check if ID exists without adding
-    func contains(_ messageID: String) -> Bool {
-        lock.lock()
-        defer { lock.unlock() }
-        return lookup.contains(messageID)
-    }
-    
-    /// Clear all entries
-    func reset() {
-        lock.lock()
-        defer { lock.unlock() }
-        
-        entries.removeAll()
-        lookup.removeAll()
-    }
-    
-    /// Periodic cleanup (called from timer)
-    func cleanup() {
-        lock.lock()
-        defer { lock.unlock() }
-        
-        cleanupOldEntries()
-        
-        // Additional memory optimization if needed
-        if entries.capacity > maxCount * 2 {
-            entries.reserveCapacity(maxCount)
-        }
-    }
-    
-    private func cleanupOldEntries() {
-        let cutoff = Date().addingTimeInterval(-maxAge)
-        
-        // Remove all entries older than cutoff
-        while let first = entries.first, first.timestamp < cutoff {
-            lookup.remove(first.messageID)
-            entries.removeFirst()
-        }
-    }
-}
-
-// MARK: - Supporting Types
-
-private struct AnnouncementPacket {
-    let nickname: String
-    let publicKey: Data
-    
-    private enum TLVType: UInt8 {
-        case nickname = 0x01
-        case noisePublicKey = 0x02
-    }
-    
-    func encode() -> Data? {
-        var data = Data()
-        
-        // TLV for nickname
-        guard let nicknameData = nickname.data(using: .utf8), nicknameData.count <= 255 else { return nil }
-        data.append(TLVType.nickname.rawValue)
-        data.append(UInt8(nicknameData.count))
-        data.append(nicknameData)
-        
-        // TLV for public key
-        guard publicKey.count <= 255 else { return nil }
-        data.append(TLVType.noisePublicKey.rawValue)
-        data.append(UInt8(publicKey.count))
-        data.append(publicKey)
-        
-        return data
-    }
-    
-    static func decode(from data: Data) -> AnnouncementPacket? {
-        var offset = 0
-        var nickname: String?
-        var publicKey: Data?
-        
-        while offset + 2 <= data.count {
-            guard let type = TLVType(rawValue: data[offset]) else { return nil }
-            offset += 1
-            
-            let length = Int(data[offset])
-            offset += 1
-            
-            guard offset + length <= data.count else { return nil }
-            let value = data[offset..<offset + length]
-            offset += length
-            
-            switch type {
-            case .nickname:
-                nickname = String(data: value, encoding: .utf8)
-            case .noisePublicKey:
-                publicKey = Data(value)
-            }
-        }
-        
-        guard let nickname = nickname, let publicKey = publicKey else { return nil }
-        return AnnouncementPacket(nickname: nickname, publicKey: publicKey)
-    }
-}
-
-private struct PrivateMessagePacket {
-    let messageID: String
-    let content: String
-    
-    private enum TLVType: UInt8 {
-        case messageID = 0x00
-        case content = 0x01
-    }
-    
-    func encode() -> Data? {
-        var data = Data()
-        
-        // TLV for messageID
-        guard let messageIDData = messageID.data(using: .utf8), messageIDData.count <= 255 else { return nil }
-        data.append(TLVType.messageID.rawValue)
-        data.append(UInt8(messageIDData.count))
-        data.append(messageIDData)
-        
-        // TLV for content
-        guard let contentData = content.data(using: .utf8), contentData.count <= 255 else { return nil }
-        data.append(TLVType.content.rawValue)
-        data.append(UInt8(contentData.count))
-        data.append(contentData)
-        
-        return data
-    }
-    
-    static func decode(from data: Data) -> PrivateMessagePacket? {
-        var offset = 0
-        var messageID: String?
-        var content: String?
-        
-        while offset + 2 <= data.count {
-            guard let type = TLVType(rawValue: data[offset]) else { return nil }
-            offset += 1
-            
-            let length = Int(data[offset])
-            offset += 1
-            
-            guard offset + length <= data.count else { return nil }
-            let value = data[offset..<offset + length]
-            offset += length
-            
-            switch type {
-            case .messageID:
-                messageID = String(data: value, encoding: .utf8)
-            case .content:
-                content = String(data: value, encoding: .utf8)
-            }
-        }
-        
-        guard let messageID = messageID, let content = content else { return nil }
-        return PrivateMessagePacket(messageID: messageID, content: content)
-    }
-}
+// Remove Nostr embedding, deduplicator, and TLV supporting types from BLE file.
